+++ Source/Game/AI/CoreAI/PetAI.cs
@@ -641,7 +641,7 @@ namespace Game.AI
             }
         }
 
-        public override void DamageTaken(Unit attacker, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null)
+        public override void DamageTaken(Unit attacker, ref double damage, DamageEffectType damageType, SpellInfo spellInfo = null)
         {
             AttackStart(attacker);
         }
